GraphQL Reactive: RequestContext cleanup
