+++ extensions/smallrye-graphql/runtime/src/main/java/io/quarkus/smallrye/graphql/runtime/spi/datafetcher/BlockingHelper.java
@@ -2,15 +2,15 @@ package io.quarkus.smallrye.graphql.runtime.spi.datafetcher;
 
 import java.util.concurrent.Callable;
 
-import graphql.schema.DataFetchingEnvironment;
-import io.quarkus.arc.InjectableContext;
-import io.quarkus.arc.ManagedContext;
 import io.smallrye.graphql.schema.model.Execute;
 import io.smallrye.graphql.schema.model.Operation;
 import io.vertx.core.Context;
 import io.vertx.core.Promise;
 
-public class BlockingHelper {
+public final class BlockingHelper {
+
+    private BlockingHelper() {
+    }
 
     public static boolean blockingShouldExecuteNonBlocking(Operation operation, Context vc) {
         // Rule is that by default this should execute blocking except if marked as non-blocking)
@@ -32,16 +32,5 @@ public class BlockingHelper {
                 future.fail(ex);
             }
         }, result);
-    }
-
-    public static void reactivate(ManagedContext requestContext, DataFetchingEnvironment dfe) {
-        if (!requestContext.isActive()) {
-            Object maybeState = dfe.getGraphQlContext().getOrDefault("state", null);
-            if (maybeState != null) {
-                requestContext.activate((InjectableContext.ContextState) maybeState);
-            } else {
-                requestContext.activate();
-            }
-        }
     }
 }
